refactor: reorganize handlers into separate files and update AppState usage
