+++ crates/forge_server/src/handlers/app_state.rs
@@ -1,0 +1,8 @@
+use std::sync::Arc;
+
+use forge_api::ForgeAPI;
+
+/// State shared between handlers
+pub struct AppState<F> {
+    pub api: Arc<ForgeAPI<F>>,
+}
